Moved Accumulator into it's own file
