+++ core/memory/index.ts
@@ -2,7 +2,7 @@
 // https://docs.google.com/spreadsheets/d/17xrEzJk5-sCB9J2mMJcVnzhbE-XH_NvczVSQH9OHvRk/edit?usp=sharing
 // Public exports for our memory "class"
 
-export { Memory, initializeCartridge, getSaveStateMemoryOffset } from './memory';
+export { Memory, initializeCartridge } from './memory';
 
 export {
   eightBitLoadFromGBMemoryWithTraps,
